file renamed from case study banner to ecosystem banner
