--- src/sections/CaseStudyBanner/index.js
@@ -1,1 +1,0 @@
-export { default as default } from './CaseStudyBanner';
